move encrypt and decrypt functions to a new module
